fix: rename schema files, restore 0171 snapshot, add trailing newline
- Rename org.ts → org-metadata.ts, org-members.ts → org-members-metadata.ts
- Update all import paths to match new file names
- Restore original 0171_snapshot.json from main (was modified during rebase)
- Fix 0172_snapshot.json prevId to match restored 0171 id
- Add trailing newline to migration SQL file
