+++ turbo/apps/web/app/api/cli/auth/test-token/route.ts
@@ -5,7 +5,7 @@ import { clerkClient } from "@clerk/nextjs/server";
 import { initServices } from "../../../../../src/lib/init-services";
 import { cliTokens } from "../../../../../src/db/schema/cli-tokens";
 import { orgCache } from "../../../../../src/db/schema/org-cache";
-import { orgMetadata } from "../../../../../src/db/schema/org";
+import { orgMetadata } from "../../../../../src/db/schema/org-metadata";
 import { orgMembersCache } from "../../../../../src/db/schema/org-members-cache";
 import { getOrgData } from "../../../../../src/lib/org/org-cache-service";
 import {
