+++ turbo/apps/web/app/api/orgs/default-agent/route.ts
@@ -4,7 +4,7 @@ import { initServices } from "../../../../src/lib/init-services";
 import { getAuthContext } from "../../../../src/lib/auth/get-auth-context";
 import { resolveOrg } from "../../../../src/lib/org/resolve-org";
 import { agentComposes } from "../../../../src/db/schema/agent-compose";
-import { orgMetadata as orgTable } from "../../../../src/db/schema/org";
+import { orgMetadata as orgTable } from "../../../../src/db/schema/org-metadata";
 import { eq, and } from "drizzle-orm";
 
 const router = tsr.router(orgDefaultAgentContract, {
