+++ turbo/apps/web/src/__tests__/api-test-helpers.ts
@@ -38,10 +38,10 @@ import type { EmailTemplate, PostSendAction } from "../lib/email/types";
 import { telegramInstallations } from "../db/schema/telegram-installation";
 import { telegramMessages } from "../db/schema/telegram-message";
 import { telegramUserLinks } from "../db/schema/telegram-user-link";
-import { orgMetadata } from "../db/schema/org";
+import { orgMetadata } from "../db/schema/org-metadata";
 import { orgCache } from "../db/schema/org-cache";
 import { orgMembersCache } from "../db/schema/org-members-cache";
-import { orgMembersMetadata } from "../db/schema/org-members";
+import { orgMembersMetadata } from "../db/schema/org-members-metadata";
 import { zeroAgents } from "../db/schema/zero-agent";
 import { userCache } from "../db/schema/user-cache";
 import { creditUsage } from "../db/schema/credit-usage";
