+++ turbo/apps/web/src/__tests__/github/api-helpers.ts
@@ -12,7 +12,7 @@ import { initServices } from "../../lib/init-services";
 import { env } from "../../env";
 import { encryptSecretValue } from "../../lib/crypto/secrets-encryption";
 import { orgCache } from "../../db/schema/org-cache";
-import { orgMetadata } from "../../db/schema/org";
+import { orgMetadata } from "../../db/schema/org-metadata";
 import {
   agentComposes,
   agentComposeVersions,
