+++ turbo/apps/web/src/db/db.ts
@@ -33,9 +33,9 @@ import * as slackOrgInstallationSchema from "./schema/slack-org-installation";
 import * as slackOrgConnectionSchema from "./schema/slack-org-connection";
 import * as slackOrgThreadSessionSchema from "./schema/slack-org-thread-session";
 import * as slackOrgPendingQuestionSchema from "./schema/slack-org-pending-question";
-import * as orgSchema from "./schema/org";
+import * as orgSchema from "./schema/org-metadata";
 import * as orgCacheSchema from "./schema/org-cache";
-import * as orgMembersSchema from "./schema/org-members";
+import * as orgMembersSchema from "./schema/org-members-metadata";
 import * as orgMembersCacheSchema from "./schema/org-members-cache";
 import * as userCacheSchema from "./schema/user-cache";
 import * as exportJobSchema from "./schema/export-job";
