+++ turbo/apps/web/src/lib/org/org-cache-service.ts
@@ -1,7 +1,7 @@
 import { eq } from "drizzle-orm";
 import { clerkClient } from "@clerk/nextjs/server";
 import { orgCache } from "../../db/schema/org-cache";
-import { orgMetadata } from "../../db/schema/org";
+import { orgMetadata } from "../../db/schema/org-metadata";
 import { logger } from "../logger";
 
 const log = logger("service:org-cache");
