+++ turbo/apps/web/src/lib/run/run-queue-service.ts
@@ -2,7 +2,7 @@ import { eq, lt, and, or, count, gt, sql, inArray } from "drizzle-orm";
 import { agentRuns } from "../../db/schema/agent-run";
 import { transitionRunStatus } from "./run-status";
 import { agentRunQueue } from "../../db/schema/agent-run-queue";
-import { orgMetadata } from "../../db/schema/org";
+import { orgMetadata } from "../../db/schema/org-metadata";
 import { env } from "../../env";
 import { logger } from "../logger";
 import { isConcurrentRunLimit } from "../errors";
